// src/components/MovieCard/types.ts
export interface IMovieCard{
    /**
     * The title of the movie
     */
    title: string;
    /**
     * the id of the movie genre to get the genre
     */
    genreId: number;
    /**
     * 
     */
    movieId: number;
    voteAverage: number;
    posterPath: string;
    sinopsis: string;
    runtime : number;
    year : number;

}